Update Style of Gem in a Box
Introduced a more readable, usable style to the application that
reduces the amount of information that repeat users need to dismiss
without removing it so far that a first time user cannot access it.

This involed making the individual gem page accessible for all gems
by a link, using icons to represent common actions, and removing the setup
instructions from every page.